+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Strip out Google redirection URL in documents.
+
 ## 0.1.0 - 2022-04-27
 
 ### Added